requirements.txt
# Combined Requirements for AI Policy Assessment Tools
# Supports both MongoDB-based compliance checker and principle-based assessment

# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
requests>=2.25.0

# AI and Machine Learning
torch>=1.9.0
transformers>=4.15.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0

# Together AI Client
together>=0.2.0

# MongoDB Integration
pymongo>=4.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Text Extraction Dependencies
opencv-python>=4.5.0
pytesseract>=0.3.8
easyocr>=1.4.0
PyPDF2>=2.0.0
pdfplumber>=0.5.0
python-docx>=0.8.11
camelot-py[cv]>=0.8.0
Pillow>=8.0.0

# Optional: For better performance and compatibility
# GPU Support (uncomment if using CUDA)
# torch>=1.9.0+cu111

# Development and Testing (optional)
# pytest>=6.0.0
# jupyter>=1.0.0
# ipykernel>=6.0.0

# Platform-specific optimizations (uncomment as needed)
# For Intel CPUs
# intel-extension-for-pytorch

# For Apple Silicon Macs
# accelerate>=0.12.0

# Additional utilities that might be helpful
tqdm>=4.60.0
python-dotenv>=0.19.0

# Optional dependencies for advanced features
faiss-cpu>=1.7.0
umap-learn>=0.5.0
plotly>=5.0.0
